Travis CI: test on a range of architectures
Travis CI supports arm64, s390x and ppc64le ("Power") now. All of
those platforms have big distros shipping Python and cryptography, so
it would be nice to test on them.

I've picked a small subset of tests to build: just Python 3.8 and the
most recent version of OpenSSL and LibreSSL. I think that provides
decent coverage without bloating the test matrix too much. (We could
also test on pypy, for example, but I'm not sure that provides enough
value to be worth the extra runtime.)

Signed-off-by: Daniel Axtens <dja@axtens.net>